Fix vggt: cast pos_embed to x.dtype to avoid float32/bfloat16 mismatch in DPTHead
make_sincos_pos_embed returns float32 (via emb.float()), which when added to
bfloat16 x upcasts the tensor to float32. The subsequent ConvTranspose2d
then fails because its weight is bfloat16 but input is float32.

+++ vggt/pytorch/loader.py
@@ -61,7 +61,7 @@ class ModelLoader(ForgeModel):
         pretrained_model_name = self._variant_config.pretrained_model_name
 
         if os.environ.get("TT_RANDOM_WEIGHTS"):
-            model = VGGT()
+            model = VGGT().float()
         else:
             model = VGGT.from_pretrained(pretrained_model_name)
